chore: 将点击触摸逻辑封装到 model 中
